fixé le temps + optimisation b&b 1

+++ held-karp.h
@@ -7,7 +7,7 @@
 using namespace std;
 
 
-double held_karp(int N, double* &graph, int pos, int visited, vector<vector<double>>& state) {
+double held_karp(int N, double* &graph, int pos, int visited, vector<vector<int>>& state) {
     if(visited == ((1 << N) - 1))
         return graph[pos*N];
